# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVG Fidget Toy Creator API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List user's projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  svgData: zod.string().describe("Raw SVG content"),
  extrudeDepth: zod.number().describe("Extrusion depth in mm"),
  keycapSize: zod.number().describe("Keycap square hole size in mm"),
  pegRadius: zod.number().describe("Circular peg radius in mm"),
  settings: zod
    .record(zod.string(), zod.unknown())
    .nullish()
    .describe("Full studio settings blob"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new fidget toy project
 */
export const createProjectBodyExtrudeDepthDefault = 4;
export const createProjectBodyKeycapSizeDefault = 14;
export const createProjectBodyPegRadiusDefault = 3.5;

export const CreateProjectBody = zod.object({
  name: zod.string(),
  svgData: zod.string(),
  extrudeDepth: zod.number().default(createProjectBodyExtrudeDepthDefault),
  keycapSize: zod.number().default(createProjectBodyKeycapSizeDefault),
  pegRadius: zod.number().default(createProjectBodyPegRadiusDefault),
  settings: zod
    .record(zod.string(), zod.unknown())
    .optional()
    .describe("Full studio settings blob"),
});

/**
 * @summary Get project statistics for the user
 */
export const GetProjectStatsResponse = zod.object({
  totalProjects: zod.number(),
  totalExports: zod.number(),
  mostRecentProject: zod
    .object({
      id: zod.number(),
      userId: zod.string(),
      name: zod.string(),
      svgData: zod.string().describe("Raw SVG content"),
      extrudeDepth: zod.number().describe("Extrusion depth in mm"),
      keycapSize: zod.number().describe("Keycap square hole size in mm"),
      pegRadius: zod.number().describe("Circular peg radius in mm"),
      settings: zod
        .record(zod.string(), zod.unknown())
        .nullish()
        .describe("Full studio settings blob"),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    })
    .nullish(),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  svgData: zod.string().describe("Raw SVG content"),
  extrudeDepth: zod.number().describe("Extrusion depth in mm"),
  keycapSize: zod.number().describe("Keycap square hole size in mm"),
  pegRadius: zod.number().describe("Circular peg radius in mm"),
  settings: zod
    .record(zod.string(), zod.unknown())
    .nullish()
    .describe("Full studio settings blob"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  name: zod.string().optional(),
  svgData: zod.string().optional(),
  extrudeDepth: zod.number().optional(),
  keycapSize: zod.number().optional(),
  pegRadius: zod.number().optional(),
  settings: zod
    .record(zod.string(), zod.unknown())
    .optional()
    .describe("Full studio settings blob"),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  name: zod.string(),
  svgData: zod.string().describe("Raw SVG content"),
  extrudeDepth: zod.number().describe("Extrusion depth in mm"),
  keycapSize: zod.number().describe("Keycap square hole size in mm"),
  pegRadius: zod.number().describe("Circular peg radius in mm"),
  settings: zod
    .record(zod.string(), zod.unknown())
    .nullish()
    .describe("Full studio settings blob"),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get the signed-in user's UI preferences
 */
export const GetUserPreferencesResponse = zod.object({
  sidebarMode: zod.enum(["simple", "advanced"]),
});

/**
 * @summary Replace the signed-in user's UI preferences
 */
export const UpdateUserPreferencesBody = zod.object({
  sidebarMode: zod.enum(["simple", "advanced"]),
});

export const UpdateUserPreferencesResponse = zod.object({
  sidebarMode: zod.enum(["simple", "advanced"]),
});