# lib/api-zod/src/generated/types/updateUserPreferencesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVG Fidget Toy Creator API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateUserPreferencesBodySidebarMode } from "./updateUserPreferencesBodySidebarMode";

export interface UpdateUserPreferencesBody {
  sidebarMode: UpdateUserPreferencesBodySidebarMode;
}