# lib/api-zod/src/generated/types/userPreferences.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SVG Fidget Toy Creator API
 * OpenAPI spec version: 0.1.0
 */
import type { UserPreferencesSidebarMode } from "./userPreferencesSidebarMode";

export interface UserPreferences {
  sidebarMode: UserPreferencesSidebarMode;
}